Separate `EssentialFeediOS` snapshots tests to the same structure as the production code (in modules) + moved snapshots
