+++ EssentialFeediOSTests/Feed UI/snapshots/FEED_WITH_CONTENT_light.png
