+++ EssentialFeediOSTests/Feed UI/snapshots/FEED_WITH_IMAGE_RETRY_dark.png
