+++ EssentialFeediOSTests/Feed UI/snapshots/FEED_WITH_IMAGE_RETRY_light.png
