+++ EssentialFeediOSTests/Image Comments UI/snapshots/IMAGE_COMMENTS_WITH_CONTENT_light.png
