+++ EssentialFeediOSTests/Shared UI/snapshots/LIST_WITH_ERROR_dark.png
