+++ EssentialFeediOSTests/Shared UI/snapshots/LIST_WITH_ERROR_light.png
